Split ecal_config_initializer path and util functions to separate file and header.
